pyproject.toml
[project]
name = "msgbusviz"
version = "0.1.0"
description = "Python client for the msgbusviz 3D message-bus visualizer"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "websocket-client>=1.7",
]

[project.optional-dependencies]
redis = ["redis>=5.0"]

[project.urls]
Repository = "https://github.com/ericfitz/msgbusviz"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["msgbusviz"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "RUF", "D"]
ignore = [
    "D100",
    "D104",
    "D105",
    "D107",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"examples/*" = ["D"]
"tests/*" = ["D"]
